Deleting logger + setting novent servlet
