+++ src/main/java/com/primeradiants/oniri/config/security/ApplicationSecurityConfig.java
@@ -26,7 +26,6 @@ public class ApplicationSecurityConfig extends WebSecurityConfigurerAdapter {
 	
 	@Autowired
 	public void configureGlobalSecurity(AuthenticationManagerBuilder auth) throws Exception {
-		System.out.println("configureGlobalSecurity");
 		//Database users retrieved with Hibernate
 		auth.userDetailsService(userDetailsService);
 		auth.authenticationProvider(authenticationProvider());
@@ -50,7 +49,6 @@ public class ApplicationSecurityConfig extends WebSecurityConfigurerAdapter {
 	@Order(1)
 	public static class ApiBasicAuthWebSecurityConfigurationAdapter extends WebSecurityConfigurerAdapter {
 		protected void configure(HttpSecurity http) throws Exception {
-			System.out.println("ApiBasicAuthWebSecurityConfigurationAdapter");
 			http
 				//TODO : determine if keeping this protection or not
 				.csrf()
